fix: anchor extraction step overlays to step root, not scrolling list
